requirements.txt
numpy==1.23.5
matplotlib==3.6.3
opencv-python==4.6.0.66
Pillow==9.3.0
tensorflow==2.12.0
scikit-learn==1.2.2
tqdm==4.64.1
